Added experts commander scaffolding.
